+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/PluginUtil.kt
@@ -125,9 +125,9 @@ internal fun Logger.logKtLintDebugMessage(
 }
 
 internal fun checkMinimalSupportedKtLintVersion(ktLintVersion: String) {
-    if (SemVer.parse(ktLintVersion) < SemVer(0, 34, 0)) {
+    if (SemVer.parse(ktLintVersion) < SemVer(0, 45, 2)) {
         throw GradleException(
-            "KtLint versions less than 0.34.0 are not supported. " +
+            "KtLint versions less than 0.45.2 are not supported. " +
                 "Detected KtLint version: $ktLintVersion."
         )
     }